move task list func
