+++ projects/valtimo/task-management/src/lib/components/task-management-column-modal/task-management-column-modal.component.ts
@@ -52,7 +52,7 @@ import {
   TaskListColumnListItem,
   TaskListColumnModalCloseEvent,
   TaskListColumnModalType,
-} from '../../models';
+} from '@valtimo/task';
 import {distinctUntilChanged} from 'rxjs/operators';
 import {isEqual} from 'lodash';
