+++ projects/valtimo/task-management/src/lib/models/index.ts
@@ -15,4 +15,3 @@
  */
 
 export * from './task-management.model';
-export * from './task-management-columns.model';
