+++ projects/valtimo/task/src/lib/components/task-list/task-list.component.html
@@ -21,7 +21,8 @@
     tasks: tasks$ | async,
     fields: fields$ | async,
     visibleTabs: visibleTabs$ | async,
-    pagination: paginationForCurrentTaskType$ | async
+    pagination: paginationForCurrentTaskType$ | async,
+    taskListColumnsForCase: taskListColumnsForCase$ | async
   } as obs"
   class="main-content"
 >
